Move observer to its own package & add wg for safety
